requirements.txt
streamlit
python-dotenv
langchain
langchain-google-vertexai
langchain-community
pypdf
chromadb
GitPython
sentence-transformers